OminousWither: bug fixes, minor changes
Fixed bug where Explosive Skull would explode at Wither's location instead of hit location
Fixed bug where Apocalypse Horsemen couldn't move due to horse movement speed being casted to int
Fixed bug where Second Phase fog did not persist with Wither unloading/reloading
Each life drain construct now makes a maximum of one sound event per drain, instead of one per entity drained. The previous behavior caused extremely loud hit sounds when attacking groups of entities

+++ OminousWither/src/main/java/io/github/poorgrammerdev/ominouswither/mechanics/ApocalypseHorsemen.java
@@ -234,8 +234,8 @@ public class ApocalypseHorsemen implements Listener {
         horse.setTamed(true);
         horse.addPotionEffect(new PotionEffect(PotionEffectType.FIRE_RESISTANCE, PotionEffect.INFINITE_DURATION, 0));
         horse.getAttribute(Attribute.GENERIC_FOLLOW_RANGE).setBaseValue(128);
-        horse.getAttribute(Attribute.GENERIC_MOVEMENT_SPEED).setBaseValue((int) this.plugin.getBossStatsManager().getStat(BossStat.APOCALYPSE_HORSE_SPEED, level, difficulty));
-        horse.getAttribute(Attribute.GENERIC_MOVEMENT_EFFICIENCY).setBaseValue((int) this.plugin.getBossStatsManager().getStat(BossStat.APOCALYPSE_HORSE_MOVEMENT_EFFICIENCY, level, difficulty));
+        horse.getAttribute(Attribute.GENERIC_MOVEMENT_SPEED).setBaseValue(this.plugin.getBossStatsManager().getStat(BossStat.APOCALYPSE_HORSE_SPEED, level, difficulty));
+        horse.getAttribute(Attribute.GENERIC_MOVEMENT_EFFICIENCY).setBaseValue(this.plugin.getBossStatsManager().getStat(BossStat.APOCALYPSE_HORSE_MOVEMENT_EFFICIENCY, level, difficulty));
     }
     
     //#region BehaviorControl
